% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{as.categorical}
\alias{as.categorical}
\title{Coerce a vector to categorical}
\usage{
as.categorical(x, ...)
}
\arguments{
\item{x}{categorical vector}

\item{...}{not used}
}
\description{
A \code{\link{categorical}} vector is an extended \code{factor} with extra attributes
}

