% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categories}
\alias{categories}
\alias{categories<-}
\alias{ncategories}
\title{Returns categories of a categorical vector}
\usage{
categories(x)

ncategories(x)

categories(x) <- value
}
\arguments{
\item{x}{\code{categorical}}

\item{value}{\code{data.frame} with information on categories. 
Each row is a category.}
}
\value{
number of categories
}
\description{
Return categories of a \code{\link{categorical}} vector

Return the number of categories

Set categories informations
}

