\name{Category-defunct}
\alias{Category-defunct}
\title{Defunct Functions in Package \pkg{Category}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{condGeneIdUniverse}
\alias{isConditional}
\alias{geneGoHyperGeoTest}
\alias{geneKeggHyperGeoTest}
\alias{cb_parse_band_hsa}
\alias{chrBandInciMat}
%
\description{
  The functions or variables listed here are no longer part of the Category package.
}
\usage{
condGeneIdUniverse()
isConditional()
geneGoHyperGeoTest()
geneKeggHyperGeoTest()
cb_parse_band_hsa()
chrBandInciMat()
}
\seealso{
  \code{\link{Defunct}}
}

