% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{is.categorical}
\alias{is.categorical}
\title{Check if a vector is a categorical vector}
\usage{
is.categorical(x)
}
\arguments{
\item{x}{categorical vector}
}
\value{
TRUE if vector is categorical
}
\description{
Check if a vector is a categorical vector
}

