% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels.categorical}
\alias{labels.categorical}
\title{return labels for the categories}
\usage{
\method{labels}{categorical}(object, labelname, ...)
}
\arguments{
\item{object}{categorical}

\item{labelname}{\code{character} name of desired label}

\item{...}{not used}
}
\value{
labelname \code{character} with label names
}
\description{
return labels for the categories
}

