% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{print.categorical}
\alias{print.categorical}
\title{Print a categorical vector}
\usage{
\method{print}{categorical}(x, ...)
}
\arguments{
\item{x}{categorical vector}

\item{...}{extra parameters, currently not used}
}
\description{
Print a \code{\link{categorical}} vector
}

