\name{categories}
\alias{categories}
\alias{categories<-}
\alias{ncategories}
\title{Returns categories}
\usage{
  categories(x)

  ncategories(x)

  categories<-(x, value)
}
\arguments{
  \item{x}{\code{categorical}}

  \item{value}{\code{data.frame} with information on
  categories}
}
\value{
  number of categories
}
\description{
  Returns categories

  Number of categories

  Set categories
}

